Cleaned Up File Structure, Refactored Codebase, and added README
