--- GPIO.c
@@ -1,76 +1,0 @@
-#include "Registers.h"
-#include "GPIO.h"
-
-unsigned char delay = 0;
-
-void portA_init(){
-	SYSCTL_RCGC2_R = 0X01;
- 	GPIO_PORTA_DIR_R = 0XFF;
- 	GPIO_PORTA_PCTL_R = (GPIO_PORTA_PCTL_R&0xFFFFFF00)+0x00000011;
-	GPIO_PORTA_LOCK_R = 0x4C4F434B;
-	delay = SYSCTL_RCGC2_R;
-  GPIO_PORTA_AFSEL_R |= 0x03;    // enable alt funct on PA1-0	GPIO_PORTA_PUR_R = 0X00;
-	GPIO_PORTA_DEN_R = 0XFF;
-	GPIO_PORTA_CR_R = 0XFF;	   
-	GPIO_PORTA_PUR_R = 0X00;	
-  GPIO_PORTA_AMSEL_R &= 0x00;  
-}
-
-
-void portF_init(){
-	volatile unsigned long delay;
-  SYSCTL_RCGC2_R |= 0x00000020;     // 1) F clock
-  delay = SYSCTL_RCGC2_R;           // delay   
-  GPIO_PORTF_LOCK_R = 0x4C4F434B;   // 2) unlock PortF PF0  
-  GPIO_PORTF_CR_R = 0x1F;           // allow changes to PF4-0       
-  GPIO_PORTF_AMSEL_R = 0x00;        // 3) disable analog function
-  GPIO_PORTF_PCTL_R = 0x00000000;   // 4) GPIO clear bit PCTL  
-  GPIO_PORTF_DIR_R = 0x0E;          // 5) PF4,PF0 input, PF3,PF2,PF1 output   
-  GPIO_PORTF_AFSEL_R = 0x00;        // 6) no alternate function
-  GPIO_PORTF_PUR_R = 0x00;          // enable pullup resistors on PF4,PF0       
-  GPIO_PORTF_DEN_R = 0x1F;          // 7) enable digital pins PF4-PF0
-}
- 
- // Initialize SysTick with busy wait running at bus clock.
-#define NVIC_ST_CTRL_R          (*((volatile unsigned long *)0xE000E010))
-#define NVIC_ST_RELOAD_R        (*((volatile unsigned long *)0xE000E014))
-#define NVIC_ST_CURRENT_R       (*((volatile unsigned long *)0xE000E018))
-#define NVIC_ST_CTRL_COUNT      0x00010000  // Count flag
-#define NVIC_ST_CTRL_CLK_SRC    0x00000004  // Clock Source 
-#define NVIC_ST_CTRL_INTEN      0x00000002  // Interrupt enable
-#define NVIC_ST_CTRL_ENABLE     0x00000001  // Counter mode
-#define NVIC_ST_RELOAD_M        0x00FFFFFF  // Counter load value
-/*
-// Initialize SysTick with busy wait running at bus clock.
-void SysTick_Init(void){
-  NVIC_ST_CTRL_R = 0;                   // disable SysTick during setup
-  NVIC_ST_RELOAD_R = NVIC_ST_RELOAD_M;  // maximum reload value
-  NVIC_ST_CURRENT_R = 0;                // any write to current clears it
-                                        // enable SysTick with core clock
-  NVIC_ST_CTRL_R = NVIC_ST_CTRL_ENABLE+NVIC_ST_CTRL_CLK_SRC;
-}*/
-// Time delay using busy wait.
-// The delay parameter is in units of the core clock. (units of 20 nsec for 50 MHz clock)
-void SysTick_Wait(unsigned long delay){
-  volatile unsigned long elapsedTime;
-  unsigned long startTime = NVIC_ST_CURRENT_R;
-  do{
-    elapsedTime = (startTime-NVIC_ST_CURRENT_R)&0x00FFFFFF;
-  }
-  while(elapsedTime <= delay);
-}
-// Time delay using busy wait.
-// This assumes 10MHz system clock.
-void SysTick_Wait1ms(unsigned long delay){
-  unsigned long i;
-  for(i=0; i<delay; i++){
-    SysTick_Wait(1000);  // wait .15s (assumes 16 MHz clock)
-  }
-}
-
-void SysTick_Wait20ms(unsigned long delay){
-  unsigned long i;
-  for(i=0; i<delay; i++){
-    SysTick_Wait(320000);  // wait .15s (assumes 16 MHz clock)
-  }
-}
