+++ README.md
@@ -1,0 +1,13 @@
+# DropTheBass
+
+TODO
+
+
+# About
+
+TODO
+
+## Setup
+
+TODO
+
