+++ drop_the_bass.d
@@ -1,2 +1,2 @@
-.\drop_the_bass.o: Drop_The_Bass.c
-.\drop_the_bass.o: UART.h
+.\drop_the_bass.o: src\Drop_The_Bass.c
+.\drop_the_bass.o: .\inc\UART.h
