+++ dtb.build_log.htm
@@ -3,52 +3,37 @@
 <pre>
 <h1>µVision Build Log</h1>
 <h2>Project:</h2>
-C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-Project File Date:  04/22/2017
+C:\Users\Michael Chalfant\ExtraDevProjects\dropthebass\dtb.uvproj
+Project File Date:  02/09/2020
 
 <h2>Output:</h2>
 Rebuild target 'Target 1'
 compiling Drop_The_Bass.c...
-UART.h(12): warning:  #1-D: last line of file ends without a newline
-  #define DEL  0x7F
-Drop_The_Bass.c(23): warning:  #1-D: last line of file ends without a newline
+src\Drop_The_Bass.c(21): warning:  #1-D: last line of file ends without a newline
   }
-Drop_The_Bass.c: 2 warnings, 0 errors
+src\Drop_The_Bass.c: 1 warning, 0 errors
 compiling EdgeTrigger.c...
-EdgeTrigger.c(86): warning:  #1-D: last line of file ends without a newline
-   */
-EdgeTrigger.c: 1 warning, 0 errors
 compiling GPIO.c...
-GPIO.h(13): warning:  #1-D: last line of file ends without a newline
-   
-GPIO.c: 1 warning, 0 errors
 compiling main.c...
-GPIO.h(13): warning:  #1-D: last line of file ends without a newline
+src\main.c(301): warning:  #1-D: last line of file ends without a newline
    
-UART.h(12): warning:  #1-D: last line of file ends without a newline
-  #define DEL  0x7F
-main.c(233): warning:  #1-D: last line of file ends without a newline
-   
-main.c: 3 warnings, 0 errors
+src\main.c: 1 warning, 0 errors
 compiling PLL.c...
-compiling ST7735.c...
-assembling Startup.s...
-compiling UART.c...
-UART.h(12): warning:  #1-D: last line of file ends without a newline
-  #define DEL  0x7F
-UART.c: 1 warning, 0 errors
 compiling PWM.c...
+compiling ST7735.c...
+compiling UART.c...
+assembling Startup.s...
 linking...
-Program Size: Code=11844 RO-data=2376 RW-data=368 ZI-data=1136  
-".\dtb.axf" - 0 Error(s), 8 Warning(s).
-"" 
---- Error: failed to execute '"" '
-"git" gui blame C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"" 
---- Error: failed to execute '"" '
-"git" gui status C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"git" gui commit C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"git" gui diffC:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"git" gui diffC:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"git" gui commit C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
-"git" gui blame C:\Users\BestBuy\Documents\CECS 490\PWM\dropthebass\dtb.uvproj
+Program Size: Code=13604 RO-data=2376 RW-data=392 ZI-data=1136  
+".\dtb.axf" - 0 Error(s), 2 Warning(s).
+
+M.c...
+compiling ST7735.c...
+compiling UART.c...
+.\inc\UART.h(12): warning:  #1-D: last line of file ends without a newline
+  #define DEL  0x7F
+src\UART.c: 1 warning, 0 errors
+assembling Startup.s...
+linking...
+Program Size: Code=13612 RO-data=2376 RW-data=416 ZI-data=1136  
+".\dtb.axf" - 0 Error(s), 5 Warning(s).
