+++ dtb.lnp
@@ -4,10 +4,10 @@
 ".\gpio.o"
 ".\main.o"
 ".\pll.o"
-".\st7735.o"
-".\startup.o"
-".\uart.o"
 ".\pwm.o"
+".\st7735.o"
+".\uart.o"
+".\startup.o"
 --ro-base 0x00000000 --entry 0x00000000 --rw-base 0x20000000 --entry Reset_Handler --first __Vectors --strict --summary_stderr --info summarysizes --map --xref --callgraph --symbols
 --info sizes --info totals --info unused --info veneers
 --list ".\dtb.map" -o .\dtb.axf
