+++ gpio.d
@@ -1,6 +1,6 @@
-.\gpio.o: GPIO.c
-.\gpio.o: Registers.h
-.\gpio.o: GPIO.h
+.\gpio.o: src\GPIO.c
+.\gpio.o: .\inc\Registers.h
+.\gpio.o: .\inc\GPIO.h
 .\gpio.o: C:\Keil\ARM\ARMCC\bin\..\include\stdio.h
 .\gpio.o: C:\Keil\ARM\ARMCC\bin\..\include\stdint.h
 .\gpio.o: C:\Keil\ARM\ARMCC\bin\..\include\stdlib.h
