+++ inc/EdgeTrigger.h
