+++ main.d
@@ -1,12 +1,13 @@
-.\main.o: main.c
+.\main.o: src\main.c
 .\main.o: C:\Keil\ARM\ARMCC\bin\..\include\stdio.h
 .\main.o: C:\Keil\ARM\ARMCC\bin\..\include\stdlib.h
-.\main.o: Registers.h
-.\main.o: PLL.h
-.\main.o: GPIO.h
+.\main.o: C:\Keil\ARM\ARMCC\bin\..\include\math.h
+.\main.o: .\inc\Registers.h
+.\main.o: .\inc\PLL.h
+.\main.o: .\inc\GPIO.h
 .\main.o: C:\Keil\ARM\ARMCC\bin\..\include\stdint.h
-.\main.o: EdgeTrigger.h
-.\main.o: PWM.h
-.\main.o: UART.h
-.\main.o: ST7735.h
-.\main.o: Drop_the_Bass.h
+.\main.o: .\inc\EdgeTrigger.h
+.\main.o: .\inc\PWM.h
+.\main.o: .\inc\UART.h
+.\main.o: .\inc\ST7735.h
+.\main.o: .\inc\Drop_the_Bass.h
