+++ src/main.c
@@ -11,32 +11,39 @@
 #include "ST7735.h"
 #include "Drop_the_Bass.h"
 
+typedef unsigned char boolean;
+#define TRUE 1
+#define FALSE 0
+
 void pitch_Handler(uint8_t string_select,uint8_t pitch_select);
 void pitch_error(unsigned long);
 void init_meter(void);
-int i ;
-volatile uint32_t timer = 0,ready = 1,CaptureData = 0,average = 0,tick = 0,averager = 0,counter = 0;
-volatile  long error = 0, control_error = 0;
-volatile int lock = 1;
+
+volatile uint32_t timer = 0, ready = 1, CaptureData = 0, average = 0,tick = 0,averager = 0,counter = 0;
+volatile long error = 0, control_error = 0;
+volatile boolean isLocked = TRUE;
 volatile long Counts = 0; 
-volatile int shiftx = 5,shifty = 0;
+volatile int shiftX = 5,shiftY = 0;
+
 unsigned int string_select = 4; //starts off on string 4
 volatile static unsigned int pitch_select[] = {0,0,0,0}; //sets pitch to standard tuning
 volatile signed int pitch1 = 0, pitch2 = 0,pitch3 = 0,pitch4 = 0;
+
 unsigned long signal_min; //minimum signal the string select will read
 unsigned long signal_max; //maximum signal the string select will read
 unsigned long pitch_index; //index for fixed signals
 unsigned int overflow; // checks if signal and string select are valid
 unsigned int timer_value;
-unsigned int count;
-unsigned int duty = 10000;
-double kp = 0.1;
+
+const unsigned int DUTY_CYCLE_VAL = 10000;
+const double ERROR_TOLERANCE = 0.1;
+const unsigned int SYSTICK_MAX_VAL = 4500;
+
 const int sting4S = 7, sting3S = 7,sting2S = 7,sting1S = 7;
 char string4[sting4S][2] = {"E","D#","D","C#","C","B"}; 
 char string3[sting3S][2] = {"A","G#","G","F#","F","E"};
 char string2[sting2S][2] = {"D","C#","C","B","A#","A"};
 char string1[sting1S][2] = {"G","F#","F","E","D#","D"};
-
  
 //updated 4/24/17
 double string4P[] = {2439,2571.2,2724.2,2886.1,3057.8,3139.6};
@@ -44,23 +51,22 @@ double string3P[] = {1818.18,1926.29,2040.8,2162.2,2290.7,2439.0};
 double string2P[] = {1362.1,1443.08,1528.9,1619.8,1716.1,1818.18};
 double string1P[] = {1020.5,1081.09,1145.3,1213.4,1285.6449,1362.1};
 
-/*
-	e,41 		| d#,38.891 | d,36.708 | C#,34.648 | c,32.703 | b,30.868
-	a,55 		| g#,51.913 | g,48.999 | f#,46.249 | f,43.654 | e,41.203
-	d,73.416|c#,69.296 	| c,65.406 | b,61.735	 | a#,58.270| a,55
-	g,97.99 | f#,92.499 | f,87.307 | e,82.407	 | d#,77.782| d,73.416
-*/
+/***********************************************************************
+	E,41 		 | D#,38.891 | D,36.708 | C#,34.648 | C,32.703 | B,30.868
+	A,55 		 | G#,51.913 | G,48.999 | F#,46.249 | F,43.654 | E,41.203
+	D,73.416 | C#,69.296 | C,65.406 | B,61.735	| A#,58.270| A,55
+	G,97.99  | F#,92.499 | F,87.307 | E,82.407	| D#,77.782| D,73.416
+************************************************************************/
 
-
-int i = 0;
- //System is running at 80Mhz  
 int main(){
 	PLL_Init();
 	UART0_Init();
 	portF_init();
 	EdgeCounter_Init();
-	PWM0A_Init(duty, 0);
-	PWM0B_Init(duty, 0);
+	
+	PWM0A_Init(DUTY_CYCLE_VAL, 0);
+	PWM0B_Init(DUTY_CYCLE_VAL, 0);
+	
 	ST7735_InitR(INITR_REDTAB);
 	ST7735_FillScreen(ST7735_BLACK);
 	ST7735_SetRotation(1);
@@ -71,11 +77,10 @@ int main(){
  	EnableInterrupts();          
 	Header();
  
-init_meter();
+  init_meter();
   while(1){
 	   GPIO_PORTF_DATA_R = 0x00;   
   pitch_error(error);
-	 
    	}		
 }
 /*
@@ -83,10 +88,12 @@ Systick timer sampling at 100KHz, 1us.
 */
 void SysTick_Handler(void){
   GPIO_PORTF_DATA_R = 0x04;        
-	lock = 0;
-	if(Counts < 4500){
+	isLocked = FALSE;
+	if(Counts < SYSTICK_MAX_VAL){
   Counts = Counts + 1;
-	}else{
+	}
+	else
+	{
 		Counts = 0; tick = 0;average = 0;averager = 0;
 		GPIO_PORTF_DATA_R = 0x00;       // disable PF2
 		NVIC_ST_CTRL_R = ~0x07; // disable SysTick with core clock and interrupts
@@ -95,8 +102,9 @@ void SysTick_Handler(void){
 /*
 Interrupt handler for pitch decrement
 */
-void GPIOPortF_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
-switch(string_select){
+void GPIOPortF_Handler(){ 		
+	GPIO_PORTF_DATA_R = 0x02;
+	switch(string_select){
 				case 4:
 					pitch4 -= 1;
  					pitch4 = (pitch4+6) % 6;
@@ -121,17 +129,17 @@ switch(string_select){
 					pitch_Handler(string_select,abs(pitch1));
 				break;
 			}
-				GPIO_PORTF_ICR_R = 0x01;
-
+	GPIO_PORTF_ICR_R = 0x01;
 }
 
 /*
 Interrupt handler for Schmitt-Trigger and RC outputs from interface buttons
 */
-void GPIOPortD_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
+void GPIOPortD_Handler(){ 		
+	GPIO_PORTF_DATA_R = 0x02;
   switch(GPIO_PORTD_RIS_R){
    case 0x01:
-      GPIO_PORTF_DATA_R  = 0x04;
+      GPIO_PORTF_DATA_R = 0x04;
  	    NVIC_ST_CTRL_R = 0x07; // enable SysTick with core clock and interrupts
   	  if(Counts>0){
 				tick++;
@@ -151,8 +159,8 @@ void GPIOPortD_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
 					 error = string1P[pitch1]-Counts;
 					 break;
 				 }
-				control_error  = error*kp; //Proportional Control System
- 				UART_OutUDec(1/(Counts*.0000001));
+				control_error  = error * ERROR_TOLERANCE; //Proportional Control System
+ 				UART_OutUDec(1/(Counts * .0000001));
 				UART_OutChar('H');
 				UART_OutChar('Z');
 				UART_OutChar(0x20);
@@ -167,8 +175,8 @@ void GPIOPortD_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
 				UART_OutChar(CR);
 				UART_OutChar(LF);
 				Counts =0;
-				}		
-       break;
+			}		
+				break;
    case 0x02:									//E-String
 		 	string_select = 4;
 	    signal_max = string4P[0]; signal_min = string4P[5];
@@ -192,22 +200,22 @@ void GPIOPortD_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
 		case 0x80:
 			switch(string_select){
 					case 1:
-					pitch1 +=1;
+					pitch1 += 1;
 					pitch1 = pitch1 % 6;
 					pitch_Handler(string_select,pitch1);
 				break;
 				case 2:
-					pitch2 +=1;
+					pitch2 += 1;
 					pitch2 = pitch2 % 6;
 					pitch_Handler(string_select,pitch2);
 				break;
 				case 3:
-					pitch3 +=1;
+					pitch3 += 1;
 					pitch3 = pitch3 % 6;
 					pitch_Handler(string_select,pitch3);
 				break;
 				case 4:
-					pitch4 +=1;
+					pitch4 += 1;
 					pitch4 = pitch4 % 6;
 					pitch_Handler(string_select,abs(pitch4));
 				break;
@@ -217,7 +225,14 @@ void GPIOPortD_Handler(){ 		GPIO_PORTF_DATA_R = 0x02;
 				GPIO_PORTD_ICR_R = 0xFF;
 }
 
-//Graphical layout for the ST7735
+/******************** ST7735 Functions **********************
+// These function controlled the Graphical Data displayed  //
+// to the user                                             // 
+// The approximate Pitch will be display at the top        //
+// Below is a tuner displayig how close to the target      // 
+// pitch this value is                                     //
+************************************************************/
+
 void pitch_Handler(uint8_t string_select,uint8_t pitch_select){
 		 switch(string_select){
 					case 1:
@@ -241,47 +256,46 @@ void pitch_Handler(uint8_t string_select,uint8_t pitch_select){
 					ST7735_DrawChar(60,0 ,string4[pitch_select][0], ST7735_GREEN , ST7735_BLACK, 8); 
 					ST7735_DrawChar(110,0 ,string4[pitch_select][1], ST7735_GREEN , ST7735_BLACK, 8); 
  				break;
-			}
-			
-		}
+			}		
+}
 	
- //Graphical layout for the ST7735
 void pitch_error(unsigned long error){
-	if((error>(0) && error <60)&&lock!=1){
+	if((error > 0 && error < 60) && isLocked != TRUE){
 			GPIO_PORTF_DATA_R = 0x08;
 			init_meter();
-			ST7735_FillRect(10+(12.5*5), 80+shifty, 14, 10, ST7735_GREEN);
-			lock = 1;
+			ST7735_FillRect(10 + (12.5*5), 80 + shiftY, 14, 10, ST7735_GREEN);
+			isLocked = TRUE;
 	 }
-	   else if((error>60 && error <2000)&&lock!= 1){
-		   init_meter();
-			 ST7735_FillRect(10+(12.5*6), 80+shifty, 14, 10, ST7735_RED);
-			 PWM0A_Duty((abs(control_error)/signal_min) * duty);
+	 else if((error > 60 && error <2000) && isLocked != TRUE){
+		 init_meter();
+		 ST7735_FillRect(10+(12.5*6), 80+shiftY, 14, 10, ST7735_RED);
+		 PWM0A_Duty((abs(control_error)/signal_min) * DUTY_CYCLE_VAL);
 
-		   lock = 1;
-				}
-				 else if((error <2000000)&&lock!= 1){
-					init_meter();
-					ST7735_FillRect(10+(12.5*4), 80+shifty, 14, 10, ST7735_RED);
-					PWM0B_Duty((abs(control_error)/signal_min) * duty);
-					lock = 1;
-				}
-  }
+		 isLocked = TRUE;
+	 }
+	 else if(error < 2000000 && isLocked != TRUE){
+		 init_meter();
+		 ST7735_FillRect(10+(12.5*4), 80+shiftY, 14, 10, ST7735_RED);
+		 PWM0B_Duty((abs(control_error)/signal_min) * DUTY_CYCLE_VAL);
+		 isLocked = TRUE;
+   }
+}
 
  void init_meter(){
-	ST7735_FillRect(10+shiftx, 80+shifty, 127, 10, ST7735_BLACK);	 
-	ST7735_DrawLine(10+shiftx,  90+shifty, 136+shiftx, 90+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(10+shiftx,  80+shifty, 136+shiftx, 80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(10+shiftx,  90+shifty, 10+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(24+shiftx,  90+shifty, 24+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(38+shiftx,  90+shifty, 38+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(52+shiftx,  90+shifty, 52+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(66+shiftx,  90+shifty, 66+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(80+shiftx,  90+shifty, 80+shiftx,  80+shifty, ST7735_YELLOW);
-  ST7735_DrawLine(94+shiftx,  90+shifty, 94+shiftx,  80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(108+shiftx, 90+shifty, 108+shiftx, 80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(122+shiftx, 90+shifty, 122+shiftx, 80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(136+shiftx, 90+shifty, 136+shiftx, 80+shifty, ST7735_YELLOW);
-	ST7735_DrawLine(136+shiftx, 90+shifty, 136+shiftx, 80+shifty, ST7735_YELLOW);
- 
+	ST7735_FillRect(10+shiftX, 80+shiftY, 127, 10, ST7735_BLACK);	 
+	ST7735_DrawLine(10+shiftX,  90+shiftY, 136+shiftX, 90+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(10+shiftX,  80+shiftY, 136+shiftX, 80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(10+shiftX,  90+shiftY, 10+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(24+shiftX,  90+shiftY, 24+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(38+shiftX,  90+shiftY, 38+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(52+shiftX,  90+shiftY, 52+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(66+shiftX,  90+shiftY, 66+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(80+shiftX,  90+shiftY, 80+shiftX,  80+shiftY, ST7735_YELLOW);
+  ST7735_DrawLine(94+shiftX,  90+shiftY, 94+shiftX,  80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(108+shiftX, 90+shiftY, 108+shiftX, 80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(122+shiftX, 90+shiftY, 122+shiftX, 80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(136+shiftX, 90+shiftY, 136+shiftX, 80+shiftY, ST7735_YELLOW);
+	ST7735_DrawLine(136+shiftX, 90+shiftY, 136+shiftX, 80+shiftY, ST7735_YELLOW);
  }
+ 
+
