+++ st7735.d
@@ -1,6 +1,6 @@
-.\st7735.o: ST7735.c
+.\st7735.o: src\ST7735.c
 .\st7735.o: C:\Keil\ARM\ARMCC\bin\..\include\stdio.h
 .\st7735.o: C:\Keil\ARM\ARMCC\bin\..\include\stdint.h
 .\st7735.o: C:\Keil\ARM\ARMCC\bin\..\include\stdlib.h
-.\st7735.o: ST7735.h
-.\st7735.o: Registers.h
+.\st7735.o: .\inc\ST7735.h
+.\st7735.o: .\inc\Registers.h
